Support quotes values in CSV files
When using the connector we ran into a few limitations of the current csv, ssv and tsv parsing that caused the connector to fail for those cases.

The specific issue we encountered was with the separator occurring in fields or headers which where then quoted to make it "legal".

An example of this is:

```
header1,header2,header3
value1,"value, 2",value3
```

Since the current parsing logic splits on all occurrences of` , `this would break, as the second line looks like it contains an additional field.

Since the project is already using jackson, which these days also supports csv parsing, it might make sense to extend the code a bit, in order to properly handle csv,ssv and tsv quoting and similar shenaningans.

This PR adds a dependency on jackson-dataformat-csv to add csv parsing functionality and replaces the current '.split(separator)' way of handling csv/ssv/tsv with the jackson parser.

We also discussed whether to keep the current behavior of trimming whitespaces around field separators during this PR. It was agreed that "proper" behavior would be to not trim whitespaces, but to keep the current behavior for compatibility purposes for now.